third_party/blink/web_tests/external/wpt/html/semantics/popups/popup-types.tentative.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="author" href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div>
  <div popup>Popup</div>
  <div popup=hint>Hint</div>
  <div popup=manual>Async</div>
  <div popup=manual>Async</div>
  <script>
  {
    const auto = document.currentScript.parentElement.querySelector('[popup=""]');
    const hint = document.currentScript.parentElement.querySelector('[popup=hint]');
    const manual = document.currentScript.parentElement.querySelectorAll('[popup=manual]')[0];
    const manual2 = document.currentScript.parentElement.querySelectorAll('[popup=manual]')[1];
    function assert_state_1(autoOpen,hintOpen,manualOpen,manual2Open) {
      assert_equals(auto.matches(':open'),autoOpen,'auto open state is incorrect');
      assert_equals(hint.matches(':open'),hintOpen,'hint open state is incorrect');
      assert_equals(manual.matches(':open'),manualOpen,'manual open state is incorrect');
      assert_equals(manual2.matches(':open'),manual2Open,'manual2 open state is incorrect');
    }
    test(() => {
      assert_state_1(false,false,false,false);
      auto.showPopUp();
      assert_state_1(true,false,false,false);
      hint.showPopUp();
      assert_state_1(true,true,false,false);
      manual.showPopUp();
      assert_state_1(true,true,true,false);
      manual2.showPopUp();
      assert_state_1(true,true,true,true);
      hint.hidePopUp();
      assert_state_1(true,false,true,true);
      auto.hidePopUp();
      assert_state_1(false,false,true,true);
      auto.showPopUp();
      hint.showPopUp();
      assert_state_1(true,true,true,true);
      auto.hidePopUp(); // Non-nested tooltips can stay open when unrelated popups are hidden.
      assert_state_1(false,true,true,true);
      hint.hidePopUp();
      manual.hidePopUp();
      assert_state_1(false,false,false,true);
      manual2.hidePopUp();
      assert_state_1(false,false,false,false);
    },'hints and manuals do not close pop-ups');

    test(() => {
      assert_state_1(false,false,false,false);
      hint.showPopUp();
      manual.showPopUp();
      manual2.showPopUp();
      assert_state_1(false,true,true,true);
      auto.showPopUp();
      assert_state_1(true,false,true,true);
      auto.hidePopUp();
      assert_state_1(false,false,true,true);
      manual.hidePopUp();
      manual2.hidePopUp();
      assert_state_1(false,false,false,false);
    },'autos close hints but not manuals');
  }
  </script>
</div>

<div>
  <div popup>Popup 1
    <div popup>Popup 2
      <p id=anchorid>Anchor</p>
      <div popup>Popup 3</div>
    </div>
  </div>
  <div popup=hint anchor=anchorid>Hint anchored to pop-up</div>
  <script>
  {
    const popUp1 = document.currentScript.parentElement.querySelectorAll('[popup=""]')[0];
    const popUp2 = document.currentScript.parentElement.querySelectorAll('[popup=""]')[1];
    const popUp3 = document.currentScript.parentElement.querySelectorAll('[popup=""]')[2];
    const hint = document.currentScript.parentElement.querySelector('[popup=hint]');
    function assert_state_2(popUp1Open,popUp2Open,popUp3Open,hintOpen) {
      assert_equals(popUp1.matches(':open'),popUp1Open,'popUp1 open state is incorrect');
      assert_equals(popUp2.matches(':open'),popUp2Open,'popup2 open state is incorrect');
      assert_equals(popUp3.matches(':open'),popUp3Open,'popup3 open state is incorrect');
      assert_equals(hint.matches(':open'),hintOpen,'hint open state is incorrect');
    }
    test(() => {
      assert_state_2(false,false,false,false);
      popUp1.showPopUp();
      popUp2.showPopUp();
      popUp3.showPopUp();
      assert_state_2(true,true,true,false);
      hint.showPopUp(); // Because hint is nested in popUp2, popUp3 should be hidden
      assert_state_2(true,true,false,true);
      popUp1.hidePopUp(); // Should close the hint, which is anchored to popUp2
      assert_state_2(false,false,false,false);
    },'hint is not closed by pre-existing auto');
  }
  </script>
</div>

<div>
  <div popup=hint>Hint
    <div popup=hint>Nested hint</div>
  </div>
  <script>
  test(() => {
    const hint1 = document.currentScript.parentElement.querySelectorAll('[popup=hint]')[0];
    const hint2 = document.currentScript.parentElement.querySelectorAll('[popup=hint]')[1];
    hint1.showPopUp();
    assert_true(hint1.matches(':open'));
    assert_false(hint2.matches(':open'));
    hint2.showPopUp();
    assert_false(hint1.matches(':open'));
    assert_true(hint2.matches(':open'));
    hint2.hidePopUp();
  },'If a popup=hint is shown, it should hide any other open popup=hint pop-ups, including ancestral pop-ups. (You can\'t nest popup=hint)');
  </script>
</div>

<div>
  <div popup="hint">Hint
    <div popup>Nested auto (note - never visible, since inside display:none subtree)</div>
  </div>
  <script>
  test(() => {
    const hint = document.currentScript.parentElement.querySelector('[popup=hint]');
    const auto = document.currentScript.parentElement.querySelector('[popup=""]');
    hint.showPopUp();
    assert_true(hint.matches(':open'));
    assert_false(auto.matches(':open'));
    auto.showPopUp();
    assert_false(hint.matches(':open'));
    assert_true(auto.matches(':open'));
    auto.hidePopUp();
  },'If a popup=auto is shown, it should hide any open popup=hint, including if the popup=hint is an ancestral pop-up of the popup=auto. (You can\'t nest a popup=auto inside a popup=hint)');
  </script>
</div>

<div>
  <div popup>Auto
    <div popup>Nested Auto</div>
    <div popup=hint>Nested hint</div>
  </div>
  <script>
  test(() => {
    const auto = document.currentScript.parentElement.querySelectorAll('[popup=""]')[0];
    const auto2 = document.currentScript.parentElement.querySelectorAll('[popup=""]')[1];
    const hint = document.currentScript.parentElement.querySelector('[popup=hint]');
    auto.showPopUp();
    auto2.showPopUp();
    assert_true(auto.matches(':open'));
    assert_true(auto2.matches(':open'));
    hint.showPopUp(); // This should hide auto2, since it is nested in auto1.
    assert_true(auto.matches(':open'));
    assert_false(auto2.matches(':open'));
    assert_true(hint.matches(':open'));
    auto.hidePopUp(); // Should hide both auto and hint.
    assert_false(auto.matches(':open'));
    assert_false(hint.matches(':open'));
  },'If you: a) show a popup=auto (call it D), then b) show a descendent popup=hint of D (call it T), then c) hide D, then T should be hidden. (A popup=hint can be nested inside a popup=auto)');
  </script>
</div>

<div>
  <div popup>Auto</div>
  <div popup=hint>Non-Nested hint</div>
  <script>
  test(() => {
    const auto = document.currentScript.parentElement.querySelector('[popup=""]');
    const hint = document.currentScript.parentElement.querySelector('[popup=hint]');
    auto.showPopUp();
    hint.showPopUp();
    assert_true(auto.matches(':open'));
    assert_true(hint.matches(':open'));
    auto.hidePopUp();
    assert_false(auto.matches(':open'));
    assert_true(hint.matches(':open'));
    hint.hidePopUp();
  },'If you: a) show a popup=auto (call it D), then b) show a non-descendent popup=hint of D (call it T), then c) hide D, then T should be left showing. (Non-nested popup=hint can stay open when unrelated popup=autos are hidden)');
  </script>
</div>
